@@ -3,19 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.2",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.0",
-    "axios": "^0.27.2",
-    "react": "^17.0.2",
+    "@reduxjs/toolkit": "^1.9.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.3.5",
+    "react": "^18.2.0",
     "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.2",
-    "react-router-dom": "^5.3.0",
-    "web-vitals": "^2.1.4"
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.10.0",
+    "web-vitals": "^3.3.1"
   },
   "devDependencies": {
+    "eslint": "^8.38.0",
+    "prettier": "^2.8.7",
     "react-scripts": "5.0.1"
   },
   "scripts": {
