@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import axios from "axios";
 import { useCookies } from "react-cookie";
-import { Redirect, useHistory, Link } from "react-router-dom";
+import { Navigate, useNavigate, Link } from "react-router-dom";
 import { Header } from "../components/Header";
 import "./signin.css";
 import { useDispatch, useSelector } from "react-redux";
@@ -12,7 +12,7 @@ import { url } from "../const";
 export const SignIn = () => {
   const auth = useSelector((state) => state.auth.isSignIn)
   const dispatch = useDispatch();
-  const history = useHistory();
+  const navigate = useNavigate();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [errorMessage, setErrorMessage] = useState();
@@ -24,14 +24,14 @@ export const SignIn = () => {
       .then((res) => {
         setCookie("token", res.data.token);
         dispatch(signIn());
-        history.push("/");
+        navigate("/");
       })
       .catch((err) => {
         setErrorMessage(`サインインに失敗しました。${err}`);
       })
   }
 
-  if(auth) return <Redirect to="/" />
+  if(auth) return <Navigate replace to="/" />
 
   return (
     <div>
